requirements.txt
azure-functions>=1.18.0
httpx>=0.24.0
xmltodict>=0.13.0